admin and user now have own controller and service
