+++ Controllers/LoginController.cs
@@ -59,50 +59,4 @@ public class LoginController : Controller
 
         return Ok(new { IsLoggedIn = false });
     }
-
-    [HttpPost("addadmin")]
-    public async Task<IActionResult> AddAdmin([FromBody] Admins admin)
-    {
-        if (admin == null) return BadRequest(new { Message = "Invalid request" });
-        if (await _loginService.addadmin(admin)) return Ok(new { Message = "Admin added" });
-        return BadRequest(new { Message = "Admin already exists" });
-    }
-
-    [HttpGet("getadmin")]
-    public async Task<IActionResult> GetAdmin()
-    {
-        var admin = await _loginService.GetAdmin();
-        return Ok(admin);
-    }
-
-    [HttpPost("deleteadmin")]
-    public async Task<IActionResult> DeleteAdmin([FromBody] Admins admin)
-    {
-        if (admin == null) return BadRequest(new { Message = "Invalid request" });
-        if (await _loginService.DeleteAdmin(admin)) return Ok(new { Message = "Admin deleted" });
-        return NotFound(new { Message = "Admin not found" });
-    }
-
-    [HttpPost("adduser")]
-    public async Task<IActionResult> AddUser([FromBody] Users user)
-    {
-        if (user == null) return BadRequest(new { Message = "Invalid request" });
-        if (await _loginService.AddUser(user)) return Ok(new { Message = "User added" });
-        return BadRequest(new { Message = "User already exists" });
-    }
-
-    [HttpGet("getuser")]
-    public async Task<IActionResult> GetUser()
-    {
-        var user = await _loginService.GetUser();
-        return Ok(user);
-    }
-
-    [HttpPost("deleteuser")]
-    public async Task<IActionResult> DeleteUser([FromBody] Users user)
-    {
-        if (user == null) return BadRequest(new { Message = "Invalid request" });
-        if (await _loginService.DeleteUser(user)) return Ok(new { Message = "User deleted" });
-        return NotFound(new { Message = "User not found" });
-    }
 }
