+++ Interfaces/IAdminService.cs
@@ -1,0 +1,6 @@
+public interface IAdminService
+{
+    Task<bool> AddAdmin(Admins admin);
+    Task<List<Admins>> GetAdmin();
+    Task<bool> DeleteAdmin(Admins admin);
+}
