+++ Interfaces/ILoginService.cs
@@ -5,10 +5,4 @@ public interface ILoginService
     Task<bool> IsSessionActive();
     Task<string> GetLoggedInUsername();
     Task<string> GetLoggedInUserRole();
-    Task<bool> addadmin(Admins admin);
-    Task<List<Admins>> GetAdmin();
-    Task<bool> DeleteAdmin(Admins admin);
-    Task<bool> AddUser(Users user);
-    Task<List<Users>> GetUser();
-    Task<bool> DeleteUser(Users user);
 }
