+++ Services/LoginService.cs
@@ -76,70 +76,4 @@ public class LoginService : ILoginService
         httpContext.Session.Remove(SessionKeyUsername);
         return true;
     }
-
-    public async Task<bool> addadmin([FromBody] Admins admin)
-    {
-        var adminToAdd = _dbContext.Admins.FirstOrDefault(a => a.Username == admin.Username);
-        if (adminToAdd != null) return false;
-
-        admin.Password = BCrypt.Net.BCrypt.HashPassword(admin.Password);
-        _dbContext.Admins.Add(admin);
-        await _dbContext.SaveChangesAsync();
-
-        return true;
-    }
-
-    public async Task<List<Admins>> GetAdmin()
-    {
-        return await _dbContext.Admins.ToListAsync();
-    }
-
-    public async Task<bool> DeleteAdmin([FromBody] Admins admin)
-    {
-        var adminToDelete = _dbContext.Admins.FirstOrDefault(a => a.Username == admin.Username);
-        if (adminToDelete == null) return false;
-
-        _dbContext.Admins.Remove(adminToDelete);
-        await _dbContext.SaveChangesAsync();
-
-        if (await IsSessionActive() && await GetLoggedInUsername() == admin.Username)
-        {
-            await logout();
-        }
-
-        return true;
-    }
-
-    public async Task<bool> AddUser([FromBody] Users user)
-    {
-        var userToAdd = _dbContext.Users.FirstOrDefault(u => u.Email == user.Email);
-        if (userToAdd != null) return false;
-
-        user.Password = BCrypt.Net.BCrypt.HashPassword(user.Password);
-        _dbContext.Users.Add(user);
-        await _dbContext.SaveChangesAsync();
-
-        return true;
-    }
-
-    public async Task<List<Users>> GetUser()
-    {
-        return await _dbContext.Users.ToListAsync();
-    }
-
-    public async Task<bool> DeleteUser([FromBody] Users user)
-    {
-        var userToDelete = _dbContext.Users.FirstOrDefault(u => u.Email == user.Email);
-        if (userToDelete == null) return false;
-
-        _dbContext.Users.Remove(userToDelete);
-        await _dbContext.SaveChangesAsync();
-
-        if (await IsSessionActive() && await GetLoggedInUsername() == user.Email)
-        {
-            await logout();
-        }
-
-        return true;
-    }
 }
